EPBDS-5620 Supported Tomcat 8 for JackRabbit repository
Tomcat 8 dissalows to read any resources outside of the application context.
So, jackrabbit properties has been moved into the *repository.war
Now bootstrap.properties and jackrabbit-repository*.xml files is located in
WEB-INF directory inside of the *repository.war artifact.

The duplicated configuration for the DEMO repository was removed because it is
unnecessary for DEMO.

production-repository was renamed to deployment-repository.