EPBDS-7884 Backed out changeset: fde1b1305896 Redundant functions
After discussing with BAs, they wanted to have such functions to prevent runtime errors.
Because of toXXX functions are tolerante to input errors, it does not need to check these inputs.
So it was decided to remove isXXX functions. Anyway if somebody will want to check format,
it is possible to achieve it using like() or isEmpty(toXXX()) functions.