EPBDS-12255 Use Spring ForwardedHeaderFilter instead of RemoteIpValve #274
It follows to RFC7239 + X-Forwarded-*

Note: There are security considerations for forwarded headers since
an application cannot know if the headers were added by a proxy,
as intended, or by a malicious client.
This is why a proxy at the boundary of trust should be configured
to remove untrusted Forwarded headers that come from the outside.