EPBDS-11381 do not remove the first slash from group name.
After investigation it was found that the keycloak has a setting in the mapper, which affects the returned group name
the full path or just the group name can be returned. This is used to be able to organize groups like top/level1/level2.
There is no point in making a custom logic for removing leading slashes, because the keycloak can add a slash in certain cases.
since this slash is not defined in specification of Saml or Cas, so we cannot make logic based on custom IDP implementation.