EPBDS-12852 EPBDS-12671 Fix automatic searching of Java to run DEMO
It was broken due incorrect determination of JRE_HOME variable.
Also fixed a case, when JRE_HOME contains a spaces.

Practically, Jetty does not require JRE_HOME for running. It just needs
a java.exe to run a server, so code for restoring JRE_HOME from the path was removed.