EPBDS-14714 Drop RMI protocol support
It challenges when requires a dynamic upgrade of rules due deserialization issue on the client side.
For the static binding it requires to write shareable Java Beans for the client & server applications.
It is not safe by default and requires to write custom code to support authorization.

In case of requirements to use RMI, it is simple to implement using:
* OpenLService for the dynamic binding
* or OpenLServiceFactoryBean for the static binding